lmkd: Fix kill reason logging in mp_event_psi()
When logging the kill reason and memory is below the high
watermark, the logic that logs the smallest watermark that
was breached always logs the min watermark, as the condition
checked is if the breached watermark is higher than the low
watermark, which will never be true in these cases. Fix the
check so that it logs the min watermark if the smallest breached
watermark is smaller than the low watermark.

CRs-Fixed: 2719960
Change-Id: I2e9d05c1f17fca177b852b3f897b1443077639b7