lmkd: Consider part of file cache as free memory
In conditions where the amount of cache used is high and the amount of
free memory is low, it is reasonable to assume a fixed part of the file
cache as free memory as some of the cache should be reclaimed fairly
soon. For the new strategy, add options to consider a given percent of
cache as free memory used during the determination of a low memory
condition.

CRs-Fixed: 3230878
Change-Id: I6491a81e47f564f5d4f00f613a3112f4fca366bb