lmkd: Consider only file cache in free memory
The current metric used for considering a part of cache as free memory
is "Cached" from meminfo. Cached considers shmem as well which could
cause applications using large amounts of shmem to not be killed in case
of high memory pressure. Change this so that only the file cache without
shmem and unevictable pages is considered.

CRs-Fixed: 3263483
Change-Id: Ifb1d502222c977848eddfa28bce5e28d0d16d239