lmkd: Fix zone parsing failing when DMA32 is empty in old strategy
in lmkd old strategy, on latest devices where DMA32 zone is empty, watermarks are parsed as 0.
The parser next expects to find nr_zone_inactive_anon, which is not
available.
So it is skipping over normal zone's watermarks till nr_zone_inactive_anon is found, ignoring Normal zone's watermarks  completely. So, overall, only DMA32's watermarks are parsed, which are 0.
This is resulting in lmkd not killing any process.

CRs-Fixed: 3372149
Change-Id: I7bdca3e4bd84bd31138f01d4501af451cbca8daf